新增PCS协议

+++ common/c_device/type_bms_i_e.go
@@ -25,7 +25,7 @@ type IBmsBasic interface {
 	GetCapacity() (uint16, error)                       // 电池容量kWh
 	GetCycleCount() (uint, error)                       // 循环次数
 
-	GetRatedPower() (float64, error)     // 额定功率， -1代表未知
+	GetRatedPower() (int32, error)       // 额定功率， -1代表未知
 	GetMaxInputPower() (float64, error)  // 最大充电功率、最大输入功率限制
 	GetMaxOutputPower() (float64, error) // 最大放电功率、最大输出功率限制
